make pv-diagrams for sim and volker data

+++ products/pvcuts/volker/volkerpvextr.py
@@ -4,7 +4,7 @@ import numpy as np
 from pvextractor import Path
 #from spectral_cube import SpectralCube
 from pvextractor import extract_pv_slice
-
+    
 import matplotlib.pyplot as plt
 from astropy.io import fits
 import os
@@ -18,11 +18,10 @@ def findpath(pp,cutlen,cutdeg):
     """pp, cutlen in pix, cutdeg in degree from north to east"""
     return ([(pp[0]-cutlen/2.*np.cos(np.deg2rad(cutdeg)),pp[1]+cutlen/2.*np.cos(np.deg2rad(cutdeg))),(pp[0]+cutlen/2.*np.cos(np.deg2rad(cutdeg)),pp[1]-cutlen/2.*np.cos(np.deg2rad(cutdeg)))])
 
-pps = [(335.62807,342.87924),(328.81196,335.5147),(321.99583,328.15016),(315.49186,320.57387),(308.12278,313.68539),(300.9139,306.6367),(294.18561,299.10743),(287.29712,291.89854),(280.72903,284.36927),(274.32113,276.67979),(267.75304,269.15052),(260.70436,262.10183),(253.65567,254.89295),(246.76719,247.68407),(239.8787,240.47519),(232.83002,233.26631),(225.94153,226.05743),(218.73265,219.00874),(211.52377,212.12026),(204.63528,204.91137),(197.90699,197.3821),(190.69811,190.33341),(183.80962,182.96433),(177.08133,175.43506),(170.35304,168.06598)]
+pps1 = [(252.1826,254.21843),(245.36649,246.85389),(238.55036,239.48935),(232.04639,231.91306),(224.67731,225.02458),(217.46843,217.97589),(210.74014,210.44662),(203.85165,203.23773),(197.28356,195.70846),(190.87566,188.01898),(184.30757,180.48971),(177.25889,173.44102),(170.2102,166.23214),(163.32172,159.02326),(156.43323,151.81438),(149.38455,144.6055),(142.49606,137.39662),(135.28718,130.34793),(128.0783,123.45945),(121.18981,116.25056),(114.46152,108.72129),(107.25264,101.6726),(100.36415,94.303516),(93.63586,86.774246),(86.90757,79.405166)]
 
-
-def makepv(name_in,makepvfits=1):
-    cutlen = 200./2.
+def makepv(name_in,pps=pps1,makepvfits=1,delpvfits=1):
+    cutlen = 200./4.
     cutdeg = 47
     name_outs = []
     for nn,pp in enumerate(pps):
@@ -30,7 +29,7 @@ def makepv(name_in,makepvfits=1):
         name_outs.append(name_out)
         if makepvfits == 1:
             pathi = findpath(pp,cutlen,cutdeg)
-            path1 = Path(pathi, width=10) # 10 pixels, 20 arcsec
+            path1 = Path(pathi, width=5) # 5 pixels, 20 arcsec
             slice1 = extract_pv_slice(name_in, path1)
             slice1.writeto(name_out,overwrite=True) 
     
@@ -52,7 +51,7 @@ def makepv(name_in,makepvfits=1):
         naxis2=mmap[0].header['naxis1']
         crval2=mmap[0].header['crval1']*3600.
         crpix2=mmap[0].header['crpix1']
-        velticks = np.arange(-2,2.,1.)
+        velticks = np.arange(-3,4.,1.)
         velpix = [(ii-crval1)/cdelt1+crpix1 for ii in velticks]
         veltickslatex = [r"$"+str(int(ii))+r"$" for ii in velticks]
         posticks = np.arange(0.,201.,20.)
@@ -78,17 +77,19 @@ def makepv(name_in,makepvfits=1):
     os.system('rm '+pdfname)
     plt.savefig(pdfname,bbox_inches='tight')
     plt.close(fig)
-    #os.system('open '+pdfname)
-    for nn,pp in enumerate(pps):
-        name_out = 'slice'+str(nn+1)+'_'+name_in
-        os.system('rm '+name_out)
+    os.system('open '+pdfname)
 
-fitsname='simc2s0p5'
-fitsname='simc2s0p2'
-fitsname='simc2s0p0'
-fitsname='simc4s0p5'
-fitsname='simc4s0p2'
-fitsname='simc2rho0p7'
-for rr in [-90,-80,-70,-60,-50,-40,-30,-20,-10,0,10,20,30,40,50,60,70,80,90]:
-    namin = fitsname+'_'+str(rr)+'.fits'
-    rp = makepv(namin)
+    if delpvfits == 1:
+        for nn,pp in enumerate(pps):
+            name_out = 'slice'+str(nn+1)+'_'+name_in
+            os.system('rm '+name_out)
+
+pps2 = [(254.23892,258.0467),(247.42281,250.68216),(240.60668,243.31762),(234.10271,235.74133),(226.73363,228.85285),(219.52475,221.80416),(212.79646,214.27489),(205.90797,207.066),(199.33988,199.53673),(192.93198,191.84725),(186.36389,184.31798),(179.31521,177.26929),(172.26652,170.06041),(165.37804,162.85153),(158.48955,155.64265),(151.44087,148.43377),(144.55238,141.22489),(137.3435,134.1762),(130.13462,127.28772),(123.24613,120.07883),(116.51784,112.54956),(109.30896,105.50087),(102.42047,98.13179),(95.692177,90.60252),(88.963887,83.23344)]
+
+fitsname='vrad_simc2s0p5_-20ccut_C18O.fits'
+fitsname='vrad_simc2s0p5_-20ccut_C18O_noisy.fits'
+fitsname='vrad_simc2s0p2_-30ccut_C18O.fits'
+fitsname='vrad_simc2s0p2_-30ccut_C18O_noisy.fits'
+rp = makepv(fitsname,pps=pps2,makepvfits=1,delpvfits=1)
+print(rp)
+
